+++ src/model_propery.h
@@ -429,7 +429,7 @@ namespace prop
 	int64_t size_bucket(int64_t n);
 	double closest_fstop(double fs);
 	int megapixels_order(double val);
-	int exp_round(double d);
+	int64_t exp_round(double d);
 
 	struct size_rounded
 	{
